requirements.txt
python-dotenv
gunicorn
redis
celery
flask
requests
beautifulsoup4
